Move helper functions to dataNormalizers file, clean up sentiment-analyses reducer
